Integrate cover into the planet
