+++ kloud51/settings.py
@@ -112,7 +112,6 @@ INSTALLED_APPS = (
     # Apps and Plugins
     'planet',
     'cmsplugin_pure_text',
-    'cmsplugin_cover',
 )
