+++ planet/models.py
@@ -199,6 +199,7 @@ class SectionItem(models.Model):
         return f'{self.title} | {self.section}'
 
 
+
 # CMS Plugin Models
 class ProductGroupPluginModel(CMSPlugin):
     """Product Group Plugin Model."""
@@ -233,3 +234,24 @@ class SectionPluginModel(CMSPlugin):
 
     def __str__(self) -> str:
         return self.section.name
+
+
+class CoverPlugin(CMSPlugin):
+    """Text Model."""
+    title = models.CharField(
+        verbose_name=_('title'),
+        max_length=250,
+        blank=True,
+        null=True,
+    )
+    sub_title = models.CharField(
+        verbose_name=_('sub title'),
+        max_length=500,
+        blank=True,
+        null=True,
+    )
+    image = FilerImageField(verbose_name=_('image'), blank=True, null=True)
+    detail = HTMLField(verbose_name=_('detail'), blank=True, null=True)
+
+    def __str__(self):
+        return self.title
